test: MMQA-2004: Migrate smoke/swap Detox tests to Appium
Move bridge, swap-action, and gasless swap smoke coverage from Detox
tests/smoke/swap into Playwright/Appium under tests/smoke-appium/swap,
with shared flow helpers and iOS Appium quote-view locator fixes.
