+++ tests/smoke-appium/swap/swap-rwa-geoblock.spec.ts
@@ -54,7 +54,7 @@ appiumTest.describe(SmokeSwap('Swap RWA Geo-block'), () => {
               type: LocalNodeType.anvil,
               options: {
                 chainId: 1,
-                loadState: './tests/smoke/swap/withTokens.json',
+                loadState: './tests/smoke-appium/swap/withTokens.json',
               },
             },
           ],
