+++ tests/smoke-appium/swap/withTokens.json
